refactor(widgets): relocate handler mixins and package dialogs/buttons
Move widgets/handlers/* into widgets/frames/app_frame/handlers/ — those
files were AppFrame mixins with no other consumers, so they now live inside
the package that composes them. Rename CardTablePanelHandler to
CardTablesHandler to disambiguate from the unrelated mixin inside the
CardTablePanel widget.

Convert every widgets/dialogs/*.py and widgets/buttons/*.py module into a
same-named package. Larger ones (feedback_dialog, guide_entry_dialog,
image_download_dialog, tutorial_dialog, deck_action_buttons) split into
dialog/panel + handlers + properties mixins. Trivial factory-style files
(help_dialog, mana_button, toolbar_buttons) wrapped as packages without a
mixin split.

Public import paths preserved via __init__.py re-exports; only AppFrame's
own imports needed to change.
