+++ widgets/buttons/deck_action_buttons/__init__.py
@@ -1,0 +1,5 @@
+"""Deck action button panel package."""
+
+from widgets.buttons.deck_action_buttons.panel import DeckActionButtons
+
+__all__ = ["DeckActionButtons"]
