+++ widgets/buttons/toolbar_buttons/__init__.py
@@ -1,0 +1,5 @@
+"""Toolbar button panel package."""
+
+from widgets.buttons.toolbar_buttons.panel import ToolbarButtons
+
+__all__ = ["ToolbarButtons"]
