+++ widgets/dialogs/feedback_dialog/__init__.py
@@ -1,0 +1,5 @@
+"""Feedback / diagnostics export dialog package."""
+
+from widgets.dialogs.feedback_dialog.dialog import FeedbackDialog, show_feedback_dialog
+
+__all__ = ["FeedbackDialog", "show_feedback_dialog"]
