+++ widgets/dialogs/guide_entry_dialog/__init__.py
@@ -1,0 +1,5 @@
+"""Guide entry editor dialog package."""
+
+from widgets.dialogs.guide_entry_dialog.dialog import GuideEntryDialog
+
+__all__ = ["GuideEntryDialog"]
