+++ widgets/dialogs/help_dialog/__init__.py
@@ -1,0 +1,5 @@
+"""Help viewer dialog package."""
+
+from widgets.dialogs.help_dialog.dialog import show_help
+
+__all__ = ["show_help"]
