+++ widgets/dialogs/image_download_dialog/__init__.py
@@ -1,0 +1,8 @@
+"""Bulk card image download dialog package."""
+
+from widgets.dialogs.image_download_dialog.dialog import (
+    ImageDownloadDialog,
+    show_image_download_dialog,
+)
+
+__all__ = ["ImageDownloadDialog", "show_image_download_dialog"]
