+++ widgets/dialogs/tutorial_dialog/__init__.py
@@ -1,0 +1,5 @@
+"""First-run tutorial wizard package."""
+
+from widgets.dialogs.tutorial_dialog.dialog import TutorialDialog, show_tutorial
+
+__all__ = ["TutorialDialog", "show_tutorial"]
